backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
google-generativeai
python-dotenv
pandas
scikit-learn
catboost
